# feed-service/docs/swagger.yaml
definitions:
  models.CommentResponse:
    properties:
      authorID:
        type: integer
      authorName:
        type: string
      content:
        type: string
      createdAt:
        type: string
      feedID:
        type: integer
      id:
        type: integer
      updatedAt:
        type: string
    type: object
  models.CreateCommentRequest:
    properties:
      content:
        type: string
      feedID:
        type: integer
    type: object
  models.CreateFeedRequest:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
  models.FeedResponse:
    properties:
      authorID:
        type: integer
      authorName:
        type: string
      commentsCount:
        type: integer
      content:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
      upvotes:
        type: integer
    type: object
  models.UpdateFeedRequest:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
info:
  contact: {}
  title: Social Feed API
  version: "1.0"
paths:
  /comment/create:
    post:
      consumes:
      - application/json
      description: Authenticated user creates a comment on a feed
      parameters:
      - description: Comment content
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CommentResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Create a new comment
      tags:
      - Comments
  /comment/paginated:
    get:
      description: Retrieve comments for a specific feed with pagination
      parameters:
      - description: Feed ID
        in: query
        name: id
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get paginated comments for a feed
      tags:
      - Comments
  /feed/:
    get:
      description: Retrieve all feeds without pagination
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.FeedResponse'
            type: array
        "500":
          description: Internal Server Error
      summary: Get all feeds
      tags:
      - Feeds
  /feed/{id}:
    delete:
      description: Delete a feed by ID
      parameters:
      - description: Feed ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Delete a feed
      tags:
      - Feeds
    get:
      description: Retrieve a single feed by its ID
      parameters:
      - description: Feed ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FeedResponse'
        "400":
          description: Bad Request
        "404":
          description: Not Found
      summary: Get feed by ID
      tags:
      - Feeds
    put:
      consumes:
      - application/json
      description: Update feed content by ID
      parameters:
      - description: Feed ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated content
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateFeedRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Update a feed
      tags:
      - Feeds
  /feed/create:
    post:
      consumes:
      - application/json
      description: Authenticated user creates a new feed
      parameters:
      - description: Feed content
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateFeedRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FeedResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Create a new feed
      tags:
      - Feeds
  /feed/paginated:
    get:
      description: Retrieve feeds with pagination parameters
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
      summary: Get paginated feeds
      tags:
      - Feeds
securityDefinitions:
  BearerAuth:
    description: Enter Bearer followed by a space and the access token from login.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"